# Speed up AOCS index build by only scan needed columns.