--- a/src/backend/access/aocs/aocsam_handler.c
+++ b/src/backend/access/aocs/aocsam_handler.c
@@ -31,6 +31,7 @@
 #include "cdb/cdbvars.h"
 #include "commands/vacuum.h"
 #include "executor/executor.h"
+#include "nodes/makefuncs.h"
 #include "nodes/nodeFuncs.h"
 #include "pgstat.h"
 #include "storage/lmgr.h"
@@ -1283,6 +1284,11 @@ aoco_index_build_range_scan(Relation heapRelation,
 	ExprContext *econtext;
 	Snapshot	snapshot;
 	bool		need_unregister_snapshot = false;
+	bool		need_create_blk_directory = false;
+	List	   *tlist = NIL;
+	List	   *qual = indexInfo->ii_Predicate;
+	Oid			blkdirrelid;
+	Oid			blkidxrelid;
 	TransactionId OldestXmin;
 
 	/*
@@ -1336,6 +1342,17 @@ aoco_index_build_range_scan(Relation heapRelation,
 	if (!IsBootstrapProcessingMode() && !indexInfo->ii_Concurrent)
 		OldestXmin = GetOldestXmin(heapRelation, PROCARRAY_FLAGS_VACUUM);
 
+	/*
+	 * If block directory is empty, it must also be built along with the index.
+	 */
+	GetAppendOnlyEntryAuxOids(RelationGetRelid(heapRelation), NULL, NULL,
+							  &blkdirrelid, &blkidxrelid, NULL, NULL);
+
+	Relation blkdir = relation_open(blkdirrelid, AccessShareLock);
+
+	need_create_blk_directory = RelationGetNumberOfBlocks(blkdir) == 0;
+	relation_close(blkdir, NoLock);
+
 	if (!scan)
 	{
 		/*
@@ -1352,12 +1369,48 @@ aoco_index_build_range_scan(Relation heapRelation,
 		else
 			snapshot = SnapshotAny;
 
-		scan = table_beginscan_strat(heapRelation,	/* relation */
-		                             snapshot,	/* snapshot */
-		                             0, /* number of keys */
-		                             NULL,	/* scan key */
-		                             true,	/* buffer access strategy OK */
-		                             allow_sync);	/* syncscan OK? */
+		/*
+		 * Scan all columns if we need to create block directory.
+		 */
+		if (need_create_blk_directory)
+		{
+			scan = table_beginscan_strat(heapRelation,	/* relation */
+										 snapshot,		/* snapshot */
+										 0,		/* number of keys */
+										 NULL,		/* scan key */
+										 true,			/* buffer access strategy OK */
+										 allow_sync);	/* syncscan OK? */
+		}
+		else
+		{
+			/*
+			 * if block directory has created, we can only scan needed column.
+			 */
+			for (int i = 0; i < indexInfo->ii_NumIndexAttrs; i++)
+			{
+				AttrNumber attrnum = indexInfo->ii_IndexAttrNumbers[i];
+				Form_pg_attribute attr = TupleDescAttr(RelationGetDescr(heapRelation), attrnum - 1);
+				Var *var = makeVar(i,
+								   attrnum,
+								   attr->atttypid,
+								   attr->atttypmod,
+								   attr->attcollation,
+								   0);
+
+				/* Build a target list from index info */
+				tlist = lappend(tlist,
+								makeTargetEntry((Expr *) var,
+												list_length(tlist) + 1,
+												NULL,
+												false));
+			}
+
+			/* Push down target list and qual to scan */
+			scan = table_beginscan_es(heapRelation,	/* relation */
+									  snapshot,		/* snapshot */
+									  tlist,		/* targetlist */
+									  qual);		/* qual */
+		}
 	}
 	else
 	{
@@ -1376,14 +1429,6 @@ aoco_index_build_range_scan(Relation heapRelation,
 	aocoscan = (AOCSScanDesc) scan;
 
 	/*
-	 * If block directory is empty, it must also be built along with the index.
-	 */
-	Oid blkdirrelid;
-	Oid blkidxrelid;
-
-	GetAppendOnlyEntryAuxOids(RelationGetRelid(aocoscan->rs_base.rs_rd), NULL, NULL,
-	                          &blkdirrelid, &blkidxrelid, NULL, NULL);
-	/*
 	 * Note that block directory is created during creation of the first
 	 * index.  If it is found empty, it means the block directory was created
 	 * by this create index transaction.  The caller (DefineIndex) must have
@@ -1392,8 +1437,7 @@ aoco_index_build_range_scan(Relation heapRelation,
 	 * blocked.  We can rest assured of exclusive access to the block
 	 * directory relation.
 	 */
-	Relation blkdir = relation_open(blkdirrelid, AccessShareLock);
-	if (RelationGetNumberOfBlocks(blkdir) == 0)
+	if (need_create_blk_directory)
 	{
 		/*
 		 * Allocate blockDirectory in scan descriptor to let the access method
@@ -1403,7 +1447,6 @@ aoco_index_build_range_scan(Relation heapRelation,
 		Assert(aocoscan->blockDirectory == NULL);
 		aocoscan->blockDirectory = palloc0(sizeof(AppendOnlyBlockDirectory));
 	}
-	relation_close(blkdir, NoLock);
 
 
 	/* GPDB_12_MERGE_FIXME */