## Speed up AOCS index build by only scan needed columns.
On first index creation (block directory is created) and only that time all columns are scanned. Later any index creation just scans needed columns.