@wangxuqi 2022-01-17
## Beach Mark
I use 1GB tpch lineitem (6 001 215 rows data) to test index build speed up.
Before this pr, index build on lineitem (l_orderkey) needs almost 4.7s 
```
create index ON lineitem (l_orderkey);
CREATE INDEX
Time: 4677.420 ms (00:04.677)
```
After this ps, index build on lineitem (l_orderkey) needs almost 1.7s
```
create index ON lineitem (l_orderkey);
CREATE INDEX
Time: 1746.630 ms (00:01.747)
```

The wider the table is, the better the performance is.